Cleaning folders and adding code for database cleaning
I rewrote the database clenaing code in R-markdown and included it in
Analyses/Emissions.Rmd. For now, this code cleans the database to leave
it as tidy data for the industries and contaminants of interest. It also
generates a plot that shows the changes in emissions for each
contaminant per industry (just for fun, I don't know if this will be
helpful).